+++ restart.sh
@@ -2,9 +2,9 @@
 
 
 rm -f *.whl *.log
-pdm build
+uv build
 cp dist/*.whl .
-pdm export --prod -o requirements.txt
+uv export --no-editable --no-emit-project -o requirements.txt
 sudo docker compose -f docker-compose.dev.yml -p dvtest stop
 sudo docker compose -f docker-compose.dev.yml -p dvtest up --build --remove-orphans -d
 sudo docker compose -f docker-compose.dev.yml -p dvtest logs -f