+++ tests/test_supabase.py
@@ -1,12 +1,16 @@
+import asyncio
+
+import pytest
 from supabase import create_client, Client
 from fastapi import Request
 from starlette.datastructures import Headers
 
 from douceville.config import config
-from douceville.auth import get_token_user
+from douceville.auth import SupabaseAuth
 
 
-def test_supabase():
+@pytest.mark.asyncio
+async def test_supabase():
     # ==========================
     # Frontend behaviour
     # ==========================
@@ -25,10 +29,16 @@ def test_supabase():
     # ==========================
     request = Request(scope={"type": "http"})
     request._headers = Headers(headers={"Authorization": f"Bearer {token}"})
-    user = get_token_user(request)
+    auth = SupabaseAuth(
+        scheme_name="scheme_name",
+        supabase_jwt_secret=config.SUPABASE_JWT_SECRET,
+        supabase_url=config.SUPABASE_URL,
+        supabase_admin_key=config.SUPABASE_ADMIN_KEY,
+    )
+    user = await auth(request)
 
     print(user)
 
 
 if __name__ == "__main__":
-    test_supabase()
+    asyncio.run(test_supabase())