+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- insertion marker -->
+## Unreleased
+
+<small>[Compare with latest](https://github.com/ydethe/douceville/compare/v1.3.11...HEAD)</small>
+
+### Added
+
+- Added API metadata ([3bbb596](https://github.com/ydethe/douceville/commit/3bbb59629bc04f4affcdd45cad5d71c701e19ae7) by Yann de The).
+
+<!-- insertion marker -->
 ## [v1.3.11](https://github.com/ydethe/douceville/releases/tag/v1.3.11) - 2024-12-11
 
 <small>[Compare with v1.3.10](https://github.com/ydethe/douceville/compare/v1.3.10...v1.3.11)</small>