+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- insertion marker -->
+## Unreleased
+
+<small>[Compare with latest](https://github.com/ydethe/douceville/compare/v1.3.12...HEAD)</small>
+
+<!-- insertion marker -->
 ## [v1.3.12](https://github.com/ydethe/douceville/releases/tag/v1.3.12) - 2024-12-12
 
 <small>[Compare with v1.3.11](https://github.com/ydethe/douceville/compare/v1.3.11...v1.3.12)</small>